Moving to STAR
test_STAR.txt contains script for testing the use of STAR.  Note that
this currently references code in Ahbi’s area.
A0 and B0 are in transition to use STAR, although D is still in the
Tophat configuration. In other words, this version is unstable and
should not be used.