Configuration
Switched to configuration files
Currently untested but designed to run test of tophat process for both
PCBC (A1) and GEO data (A2).  Also changed B0 and B1 so that they don't
need to know whether we are using tophat or STAR (that detail is handled
by D and in the configuration file).