Merge Provenance with Script
Pipeline script and corresponding provenance files have been merged so
that provenance is recorded as the files are being made.  (Provenance
scripts are now superfluous.) Currently testing.
Note that this version has some items disabled for the sake of speed in
testing.  In particular, C is disabled within B. (i.e. we do not
currently convert sra to fastq).  This is a simple matter of removing a
comment character but needs to be noted.